+++ CHANGES.md
@@ -5,6 +5,8 @@
 
 - `Interval.flush` will not reset if canceled.
 - `SizeUtils.formatSize` support `fixedPoint` notation.
+- `ListUtils.indexBy` -> `ListUtils.indexToMap`, and add a `ListUtils.indexToObject`.
+- `ListUtils.groupBy` -> `ListUtils.groupToMap`, and add a `ListUtils.groupToObject`.
 
 
 # 2.0.1